v0.13
Took 6 hours 27 minutes

+++ src/reengineering/Bishop.java
@@ -19,7 +19,7 @@ public class Bishop extends APiece {
      * @param sY souradnice stredove souradnice Y
      * @param isWhite test na barvu figurky (bila/cerna)
      */
-    public Bishop(int sX, int sY, boolean isWhite) {
+    public Bishop(double sX, double sY, boolean isWhite) {
         super(sX, sY, isWhite);
     }
